Move component-proxy to components folder
